Reintroduce the stage to distribute stake across selected AVSs
